Hide/show candidate window via generic option
Adding a new option to "m17nlib.scm" to add new option is
still possible, however we just need to use the known option.
This may cause confusion, because "Other input methods" (that
is that name of "Generic") often excludes "m17n-lib". So
it is a feature (possibly a bug) that "m17n-lib" uses a
generic option for its. This also requires dependencies,
and we will not be able to have a separated settings for "m
17n-lib". This is almost good, bc. showing/hiding the window
for candidates may be a personal setting: someone just likes
to see them all (for any input methods/settings), while some
others just like to see nothing