Remove :state icon on systray
The systray has two groups (aka :branches) of icons on systray.
One is to list all IMs, the other is to indicate the state of the
current IM (on, off, or direct_input). There may be some other
group but I do not know.

Because we tell :uim to switch to :direct_input as an alternative
of :toggle action, we do not really need a second icon to show
the same thing. We just remove them, and this will make the icon
bigger/viewable on the systray. The combination "^ Alt \" is
still valid: it can be used to switch :telex <-> :vni. Note that
some user option may make new feature buggy. Also note that we
only test this feature on :Gtk2, but :Gtk3 should work, too.

I spent my two days (sat & sun) to figure out how I can patch.
I first thought there are some Gtk code that creates two groups.
Unfortunately that is completely wrong. The :toolbar reads the
instructions from :uim_fd (file descriptor that contains control
messages), and parses the messages to create buttons and icons
for the systray. See (helper_toolbar_prop_list_update). That
is why you see lines that use (strcmp) to check if the current
message/line wants to create a :branch (:group) or :leaf (icon)