qt5.im.input_context: check if #widget > 0 before #unsetFocus()
Due to our patch 512a3c879de680 (theslinux bugs#138), #uim may invoke
#im-focus-out-handler twice [!?]. The 2nd call will apply on an empty widget,
and in #QUimPlatformInputContext::unsetFocus (qt5), a call to
  uim_focus_out_context(m_uc);
will make application crashed. The crash can be preprocedured as below

* Open #texmaker (qt5)
* Type something and in pre-edit phase
* Move out the mouse (focus lost)
* Enjoy the crash

This not happens on #qt4, though.

The good news is that #qt5 still holds the pre-edit window. We end up in
non-persistent behavior

* Qt4: when focus is lost, commit (or cancel) happens
* Qt5: when focus is lost, context is hold

TODO: explain why the problem does not happens on #qt4