!bug: Use #QT5_CANDWIN to remove all #candidate support
When #nocandwin is used (--with-qt5 is not specified), the #module still
calls to #candidate window support, and that creates zoombies processes
for every commit string. When --with-qt5 is enabled, #candidate support
is too crazy, because #/usr/lib/uim/uim-candwin-qt5 will call to itself
and generate a lot of child processes. System will be quickly full of Z.